Function to hold common parameters.

Function to hold common parameters.

common_params(
  experiment,
  genome_assembly,
  genome_annotation,
  sample_sheet,
  threshold,
  samples,
  use_normalized,
  ncol,
  log2fc_cutoff,
  fdr_cutoff,
  dominant,
  exclude_antisense,
  data_conditions,
  rasterize,
  raster_dpi,
  return_table
)

Arguments

experiment
TSRexploreR object.

genome_assembly
Genome assembly in FASTA or BSgenome format.

genome_annotation
Genome annotation in GTF, GFF3, or TxDb format.

sample_sheet
A sample sheet data.frame or tab delimited file. Must have the columns 'sample_name', 'file_1', and 'file_2'. Additional meta-data columns can be added with sample information such as condition and batch.

threshold
TSSs or TSRs with a score below this value will not be considered.

samples
A vector of sample names to analyze.

use_normalized
Whether to use the normalized (TRUE) or raw (FALSE) counts.

ncol
Integer specifying the number of columns to arrange multiple plots.

log2fc_cutoff
Differential features not meeting this |log2(fold change)| threshold will not be considered.

fdr_cutoff
Differential features not meeting this significance threshold will not be considered.

dominant
If TRUE, will only consider the highest-scoring TSS per gene, transcript, or TSR or highest-scoring TSR per gene or transcript.

exclude_antisense
Remove antisense TSSs/TSRs prior to analysis.

data_conditions
Apply advanced conditions to the data.

rasterize
Rasterize a ggplot.

raster_dpi
If rasterization is set, this controls the rasterization DPI.

return_table
Return a table of results instead of a plot.